pyproject.toml
[tool.poetry]
name = "sailinganalysis"
version = "0.1.0"
description = ""
authors = ["Felipe Sodré M. Barros"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
contextily = "^1.3.0"
plotly = "^5.14.1"
datetime = "^5.1"
fiona = "^1.9.3"
geopandas = "^0.12.2"
movingpandas = "^0.15rc1"
python-dotenv = "^1.0.0"
meteostat = "^1.6.5"
sqlalchemy = "^2.0.9"
alembic = "^1.10.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
